# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AcctMarket - Social Media Account Selling Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/change-password:
    post:
      operationId: changePassword
      tags: [users]
      summary: Change user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # CATEGORIES
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products (with optional filters)
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: quality
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted

  # ORDERS
...
  /wallet/transactions:
    get:
      operationId: listTransactions
      tags: [wallet]
      summary: List user transactions
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"

  # DEPOSITS
...
  /support/tickets:
    get:
      operationId: listTickets
      tags: [support]
      summary: List user support tickets
      responses:
        "200":
          description: User tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [support]
      summary: Create support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketBody"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /support/tickets/{id}:
    get:
      operationId: getTicket
      tags: [support]
      summary: Get ticket with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketWithMessages"
  /support/tickets/{id}/reply:
    post:
      operationId: replyTicket
      tags: [support]
      summary: Reply to support ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyTicketBody"
      responses:
        "201":
          description: Reply sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketMessage"

  # SETTINGS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        balance:
          type: number
        isAdmin:
          type: boolean
        isBanned:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - email
        - fullName
        - balance
        - isAdmin
        - isBanned
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        username:
          type: string
        fullName:
          type: string
        bio:
          type: string
        avatar:
          type: string
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword
...
    Product:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        quality:
          type: string
        stockCount:
          type: integer
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        previewInfo:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        totalSold:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - categoryId
        - categoryName
        - name
        - slug
        - price
        - quality
        - stockCount
        - isAvailable
        - isFeatured
        - totalSold
        - createdAt
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - totalPages
    CreateProductBody:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        quality:
          type: string
        stockLogs:
          type: string
        previewInfo:
          type: string
        imageUrl:
          type: string
        isFeatured:
          type: boolean
        isAvailable:
          type: boolean
      required:
        - categoryId
        - name
        - slug
        - price
        - quality
    UpdateProductBody:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        quality:
          type: string
        stockLogs:
          type: string
        previewInfo:
          type: string
        imageUrl:
          type: string
        isFeatured:
          type: boolean
        isAvailable:
          type: boolean
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        amount:
          type: number
        description:
          type: string
        balanceBefore:
          type: number
        balanceAfter:
          type: number
        createdAt:
          type: string
      required:
        - id
        - type
        - amount
        - description
        - balanceBefore
        - balanceAfter
        - createdAt
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - transactions
        - total
        - page
        - totalPages
...
    Ticket:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        subject:
          type: string
        status:
          type: string
        priority:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - subject
        - status
        - priority
        - createdAt
        - updatedAt
    TicketMessage:
      type: object
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        userId:
          type: integer
        message:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - ticketId
        - userId
        - message
        - isAdmin
        - createdAt
    TicketWithMessages:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        subject:
          type: string
        status:
          type: string
        priority:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/TicketMessage"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - subject
        - status
        - priority
        - messages
        - createdAt
        - updatedAt
    CreateTicketBody:
      type: object
      properties:
        subject:
          type: string
        message:
          type: string
        priority:
          type: string
      required:
        - subject
        - message
    ReplyTicketBody:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...