# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AcctMarket - Social Media Account Selling Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  categoryId: number;
  categoryName: string;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  quality: string;
  stockCount: number;
  isAvailable: boolean;
  isFeatured: boolean;
  /** @nullable */
  previewInfo?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  totalSold: number;
  createdAt: string;
}